Fix cross-compiling to .Net 462
In order to deliver a package that supports full framework Windows PowerShell,
I've had to cave a bit and reference a different package for PowerShell
and to make a minor concession on error handling regarding the script stack trace.

To make it possible to deliver a release package, I'm including an Install.ps1 script.
Eventually I'll ship it all to Chocolatey...